pyproject.toml
[project]
name = "tableau-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.13",
    "fastmcp>=2.8.1",
    "mcp[cli]>=1.9.4",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
]
